feat: move budget page into new budget component
