+++ cash-map/src/app/app.component.html
@@ -1,69 +1,1 @@
-<div class="page">
-  <h1 class="title">Cash Map</h1>
-  <app-cache-controls [(budget)]="budget"></app-cache-controls>
-
-  <div class="sections">
-    <div class="section">
-      <h3>Income</h3>
-      <div class="fields">
-        <mat-form-field appearance="fill">
-          <mat-label>Income (per year)</mat-label>
-          <span matPrefix>$</span>
-          <input matInput type="number" min="0" [(ngModel)]="budget.income" />
-        </mat-form-field>
-        <mat-form-field appearance="fill">
-          <mat-label>Percentage of income paid in taxes</mat-label>
-          <input matInput type="number" min="0" [(ngModel)]="budget.percentageOfIncomeForTaxes" />
-          <span matSuffix>%</span>
-        </mat-form-field>
-      </div>
-    </div>
-
-    <div class="section">
-      <h3>Expenses (recurring monthly)</h3>
-      <button class="add-button" mat-raised-button color="primary" (click)="addMonthlyExpense()">Add</button>
-      <div class="expenses">
-        <div class="expense" *ngFor="let expense of monthlyExpenses()">
-          <mat-form-field appearance="fill">
-            <mat-label>Label</mat-label>
-            <input matInput type="text" placeholder="Label" [(ngModel)]="expense.label" />
-          </mat-form-field>
-          <mat-form-field appearance="fill">
-            <mat-label>Amount</mat-label>
-            <span matPrefix>$</span>
-            <input matInput type="number" min="0" step="0.01" placeholder="Amount" [(ngModel)]="expense.value" />
-          </mat-form-field>
-          <button mat-icon-button color="accent" (click)="removeExpense(expense.id)">
-            <mat-icon aria-hidden="false" aria-label="Delete">delete</mat-icon>
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <div class="section">
-      <h3>Non-recurring Expenses</h3>
-      <button class="add-button" mat-raised-button color="primary" (click)="addNonRecurringExpense()">Add</button>
-      <div class="expenses">
-        <div class="expense" *ngFor="let expense of nonRecurringExpenses()">
-          <mat-form-field appearance="fill">
-            <mat-label>Label</mat-label>
-            <input matInput type="text" placeholder="Label" [(ngModel)]="expense.label" />
-          </mat-form-field>
-          <mat-form-field appearance="fill">
-            <mat-label>Amount</mat-label>
-            <span matPrefix>$</span>
-            <input matInput type="number" min="0" step="0.01" placeholder="Amount" [(ngModel)]="expense.value" />
-          </mat-form-field>
-          <button mat-icon-button color="accent" (click)="removeExpense(expense.id)">
-            <mat-icon aria-hidden="false" aria-label="Delete">delete</mat-icon>
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <div class="section">
-      <h3>Net (annually)</h3>
-      <p>${{this.net() | number:'1.2-2'}}</p>
-    </div>
-  </div>
-</div>
+<app-budget></app-budget>
