+++ cash-map/src/app/app.module.ts
@@ -8,12 +8,14 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MaterialModule } from './material.module';
 import { ClearCacheDialogComponent } from './clear-cache-dialog/clear-cache-dialog.component';
 import { CacheControlsComponent } from './cache-controls/cache-controls.component';
+import { BudgetComponent } from './budget/budget.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     ClearCacheDialogComponent,
-    CacheControlsComponent
+    CacheControlsComponent,
+    BudgetComponent
   ],
   imports: [
     BrowserModule,
